Separate parser building from cli processing args
